Add image generation based off of taste profile and refactor
